requirements_new.txt
# Google API クライアント
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# データ処理
python-dateutil>=2.8.2
